# Rewrite Integer#times with Ruby