For maintenance reasons and [future speed-up](https://speakerdeck.com/k0kubun/the-method-jit-compiler?slide=150) with (not-committed-yet) yield inlining by JIT.

## Interpreter

```
$ benchmark-driver benchmark/loop_times.yml -v --rbenv 'before::yjit-release-before;after::yjit-release-after' --repeat-count=4
before: ruby 3.2.0dev (2022-10-21T21:06:34Z master 87bb0bee6b) [x86_64-linux]
after: ruby 3.2.0dev (2022-10-21T21:21:05Z builtin-times 165efec492) [x86_64-linux]
Warming up --------------------------------------
             10.times    11.241M i/s -     11.473M times in 1.020624s (88.96ns/i, 195clocks/i)
        10.times{|e|}     4.647M i/s -      4.647M times in 1.000029s (215.20ns/i, 473clocks/i)
      1000.times{|e|}    52.275k i/s -     56.980k times in 1.089995s (19.13μs/i)
30_000_000.times{|e|}      1.801 i/s -       2.000 times in 1.110794s (555.40ms/i)
Calculating -------------------------------------
                          before       after
             10.times    13.312M     12.616M i/s -     33.724M times in 2.533426s 2.673206s
        10.times{|e|}     4.987M      3.376M i/s -     13.940M times in 2.795105s 4.128913s
      1000.times{|e|}    55.811k     37.200k i/s -    156.826k times in 2.809951s 4.215721s
30_000_000.times{|e|}      1.857       1.266 i/s -       5.000 times in 2.692246s 3.950562s

Comparison:
                          10.times
               before:  13311817.6 i/s
                after:  12615752.3 i/s - 1.06x  slower

                     10.times{|e|}
               before:   4987373.4 i/s
                after:   3376246.8 i/s - 1.48x  slower

                   1000.times{|e|}
               before:     55810.9 i/s
                after:     37200.3 i/s - 1.50x  slower

             30_000_000.times{|e|}
               before:         1.9 i/s
                after:         1.3 i/s - 1.47x  slower
```