@k0kubun 2022-10-21
I benchmarked it again and updated the description. For some reason, the result is significantly different today and I'm no longer comfortable merging it. I'll revisit this after MJIT or YJIT gets `yield` inlining.