@k0kubun 2022-11-01
Now that we at least have https://github.com/ruby/ruby/pull/6640, I'll play with this branch a bit again.