@k0kubun 2022-11-03
I'll revisit this a bit differently.